Extract class InputParser
